# frontend/src/client/openapi/workspace.openapi.json
{
  "components": {
    "schemas": {
...
      "AgentTaskRequest": {
        "properties": {
          "context_file_ids": {
            "items": {
              "type": "string"
            },
            "title": "Context File Ids",
            "type": "array"
          },
          "kb_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kb Id"
          },
          "task": {
            "minLength": 1,
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "task"
        ],
        "title": "AgentTaskRequest",
        "type": "object"
      },
      "AgentTaskResponse": {
        "properties": {
          "final_answer": {
            "title": "Final Answer",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/AgentStep"
            },
            "title": "Steps",
            "type": "array"
          },
          "task": {
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task",
          "status",
          "steps",
          "final_answer"
        ],
        "title": "AgentTaskResponse",
        "type": "object"
      },
...
      "FolderCreate": {
        "properties": {
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "scope": {
            "default": "personal",
            "enum": [
              "personal",
              "team"
            ],
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "FolderCreate",
        "type": "object"
      },
      "FolderItem": {
        "properties": {
          "children": {
            "items": {
              "$ref": "#/components/schemas/FolderItem"
            },
            "title": "Children",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "permission": {
            "title": "Permission",
            "type": "string"
          },
          "scope": {
            "enum": [
              "personal",
              "team"
            ],
            "title": "Scope",
            "type": "string"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          }
        },
        "required": [
          "id",
          "name",
          "scope",
          "permission"
        ],
        "title": "FolderItem",
        "type": "object"
      },
      "FolderTreeResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FolderItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "FolderTreeResponse",
        "type": "object"
      },
      "FolderUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "title": "FolderUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "KnowledgeBaseCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "KnowledgeBaseCreate",
        "type": "object"
      },
      "KnowledgeBaseListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeBasePublic"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "KnowledgeBaseListResponse",
        "type": "object"
      },
      "KnowledgeBasePublic": {
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "document_count": {
            "title": "Document Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "status",
          "document_count",
          "chunk_count",
          "updated_at"
        ],
        "title": "KnowledgeBasePublic",
        "type": "object"
      },
      "KnowledgeBaseUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "active",
                  "archived"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "KnowledgeBaseUpdate",
        "type": "object"
      },
      "KnowledgeDocumentCreate": {
        "properties": {
          "file_id": {
            "minLength": 1,
            "title": "File Id",
            "type": "string"
          }
        },
        "required": [
          "file_id"
        ],
        "title": "KnowledgeDocumentCreate",
        "type": "object"
      },
      "KnowledgeDocumentListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeDocumentPublic"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "KnowledgeDocumentListResponse",
        "type": "object"
      },
      "KnowledgeDocumentPublic": {
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "file_id": {
            "title": "File Id",
            "type": "string"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "index_status": {
            "enum": [
              "queued",
              "indexing",
              "indexed",
              "failed"
            ],
            "title": "Index Status",
            "type": "string"
          },
          "kb_id": {
            "title": "Kb Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kb_id",
          "file_id",
          "file_name",
          "index_status",
          "chunk_count",
          "updated_at"
        ],
        "title": "KnowledgeDocumentPublic",
        "type": "object"
      },
...
      "QARequest": {
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "kb_id": {
            "title": "Kb Id",
            "type": "string"
          },
          "question": {
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "stream": {
            "default": false,
            "title": "Stream",
            "type": "boolean"
          },
          "top_k": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "kb_id",
          "question"
        ],
        "title": "QARequest",
        "type": "object"
      },
      "QAResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "message_id": {
            "title": "Message Id",
            "type": "string"
          }
        },
        "required": [
          "conversation_id",
          "message_id",
          "answer",
          "citations"
        ],
        "title": "QAResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WHU API v2",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v2/agents/tasks": {
      "post": {
        "operationId": "create_agent_task_api_v2_agents_tasks_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Agent Task"
      }
    },
...
    "/api/v2/folders": {
      "post": {
        "operationId": "create_folder_api_v2_folders_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Folder"
      }
    },
    "/api/v2/folders/tree": {
      "get": {
        "operationId": "folder_tree_api_v2_folders_tree_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderTreeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Folder Tree"
      }
    },
    "/api/v2/folders/{folder_id}": {
      "delete": {
        "operationId": "delete_folder_api_v2_folders__folder_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "folder_id",
            "required": true,
            "schema": {
              "title": "Folder Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Folder"
      },
      "patch": {
        "operationId": "update_folder_api_v2_folders__folder_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "folder_id",
            "required": true,
            "schema": {
              "title": "Folder Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Folder"
      }
    },
...
    "/api/v2/knowledge-bases": {
      "get": {
        "operationId": "list_knowledge_bases_api_v2_knowledge_bases_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Knowledge Bases"
      },
      "post": {
        "operationId": "create_knowledge_base_api_v2_knowledge_bases_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBasePublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Knowledge Base"
      }
    },
    "/api/v2/knowledge-bases/{kb_id}": {
      "patch": {
        "operationId": "update_knowledge_base_api_v2_knowledge_bases__kb_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBasePublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Knowledge Base"
      }
    },
    "/api/v2/knowledge-bases/{kb_id}/documents": {
      "get": {
        "operationId": "knowledge_documents_api_v2_knowledge_bases__kb_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeDocumentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Knowledge Documents"
      },
      "post": {
        "operationId": "add_knowledge_document_api_v2_knowledge_bases__kb_id__documents_post",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeDocumentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeDocumentPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Knowledge Document"
      }
    },
...
    "/api/v2/qa/query": {
      "post": {
        "operationId": "qa_query_api_v2_qa_query_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QARequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QAResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Qa Query"
      }
    },
...
    "/api/v2/tools": {
      "get": {
        "operationId": "tools_api_v2_tools_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tools"
      }
    },
...
    "/api/v2/workspace/snapshot": {
      "get": {
        "operationId": "workspace_snapshot_api_v2_workspace_snapshot_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workspace Snapshot"
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}